Python Bash Script
This is a python program designed to run as a bash script using the same approach Nate used from the Jupyter Notebook. I was using Spyder as my development platform from within the anaconda navigator. When you run the algorithm type 'python name_of_program' and then pass it a txt file with all of your keys. This is a very rough v 0.2 and still needs to connect to a steady http: address to work